requirements.txt
appdirs==1.4.4
asttokens==2.0.8
asyncio-nats-client==0.11.4
backcall==0.2.0
bcrypt==4.0.1
black==21.7b0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.1
contourpy==1.0.5
cryptography==38.0.4
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
distro==1.4.0
distro-info==1.0
docker==6.0.1
ed25519==1.5
entrypoints==0.4
executing==1.1.1
FileIO==0.3
fonttools==4.37.4
idna==2.8
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.4.2
kiwisolver==1.4.4
matplotlib==3.6.1
matplotlib-inline==0.1.6
modbus-tk==1.1.2
mypy==0.910
mypy-extensions==0.4.3
nest-asyncio==1.5.6
nkeys==0.1.0
numpy==1.21.2
packaging==21.3
paramiko==2.12.0
parso==0.8.3
pathspec==0.9.0
persist-queue==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.4.0
prompt-toolkit==3.0.31
protobuf==3.17.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.21.0
pycparser==2.21
Pygments==2.13.0
PyGObject==3.36.0
pykka==3.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-snappy==0.6.0
pyzmq==24.0.1
regex==2021.7.6
requests==2.28.1
requests-unixsocket==0.2.0
Rx==3.2.0
six==1.16.0
ssh-import-id==5.10
stack-data==0.5.1
toml==0.10.2
tomli==1.0.4
tornado==6.2
traitlets==5.4.0
typing-extensions==3.10.0.0
unattended-upgrades==0.1
urllib3==1.26.13
wcwidth==0.2.5
websocket-client==1.4.2
